pyproject.toml
[project]
name = "hibikasu-prd-reviewer"
version = "0.1.0"
description = "AI Review Panel system using Google ADK for PRD review by specialist agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "toml>=0.10.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.5.0",
    "ruff>=0.4.0",
    "bandit>=1.7.0",
    "pip-audit>=2.6.0",
    "pre-commit>=3.0.0",
]

docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "myst-parser>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hibikasu_agent"]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
exclude = [".venv"]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "PL",     # Pylint
]
ignore = [
    "G004",
    "F821",
    "C401",
    "C408",
    "RUF001",
    "RUF002",
    "RUF003",
    "ANN101",
    "ANN102",
    "PERF203",
    "PERF401",
    "PLR2004",
    "PLC2401",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401", "E402", "PLR2004", "PLR0913"]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101", "B311", "B404", "B603"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov-branch",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.8.0",
    "urllib3>=2.5.0",
]
